chg int to decimal

+++ db/migrate/20171024024916_change_price_to_decimal.rb
@@ -1,5 +1,5 @@
 class ChangePriceToDecimal < ActiveRecord::Migration[5.1]
   	def change
-		change_column :products, :price, :int
+		change_column :products, :price, :decimal
 	end
 end
